Fix header to be the same as cpp

+++ button.h
@@ -2,10 +2,12 @@
 #ifndef BUTTON_H
 #define BUTTON_H
 
+#include <Arduino.h>
+
 const unsigned char BUTTON_NONE = 100;
 
-bool button_isPressed(int buttonNumber);
-void button_wait(int buttonNumber);
+bool button_isPressed(uint8_t buttonNumber);
+void button_wait(uint8_t buttonNumber);
 void button_ISR();
 
 #endif
